feat: #109
Current buffer is now attached whenever `setup` is called.

fix: #121

Lazy loading issue resolved.

fix: #122

Submodule utils is now exported from vgit and updated README.md.

fix: #123

internal hunk generation is now conditional based on vim.diff being
available.
