+++ lua/vgit/decorators/VirtualLineNrDecorator.lua
@@ -40,7 +40,7 @@ function VirtualLineNrDecorator:mount()
     events.buf.on(
         self.content_buf,
         'WinClosed',
-        string.format(':lua require("vgit").renderer.hide_windows({ %s })', self.win_id),
+        string.format(':lua _G.package.loaded.vgit.renderer.hide_windows({ %s })', self.win_id),
         { once = true }
     )
     return self
